RevitLintelsManager: Создание класса команды менеджера

+++ src/RevitLintelsManager/Models/Rules/LintelRule.cs
@@ -1,8 +1,8 @@
 ﻿namespace RevitLintelsManager.Models;
 
 public class LintelRule {
-    public double MinWidth { get; set; }
-    public double MaxWidth { get; set; }
-    public double WallSupport { get; set; }
+    public int MinWidth { get; set; }
+    public int MaxWidth { get; set; }
+    public int WallSupport { get; set; }
     public string LintelType { get; set; }
 }
